{{cookiecutter.project_repository_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.python_package_name}}"
version = "{{cookiecutter.package_version}}"
description = ""
authors = ["{{cookiecutter.author}} <{{cookiecutter.author_email}}>"]

[tool.poetry.dependencies]
python = ">={{cookiecutter.minimum_python_version}},<3.12"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
bump2version = "^1.0.1"
cookiecutter = "^2.1.1"
coverage = "^7.0.1"
coverage-badge = "^1.1.0"
ipykernel = "^6.19.4"
isort = "^5.11.4"
nb-black-only = ">=1.0.9"
pre-commit = "^2.21.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-html = "^3.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3

[coverage.paths]
source = "{{cookiecutter.python_package_name}}"

[coverage.run]
branch = true

[coverage.report]
fail_under = 10
show_missing = true
